minute_of_day was wrong for 99.7% of rows since day one
polars keeps .dt.hour()'s native Int8 through a multiply, so hour*60
overflows above 127 and wraps. 21:10 was being encoded as -10 rather than
1270. Every model in this project trained on it, and nothing broke
loudly because the result is still locally monotonic inside most hours --
it is a wrapped transform, not noise, which is exactly why it survived
seven submissions unnoticed.

Found by an agent auditing feature construction rather than by anything
we measured. Fixing it is guaranteed safe: a broken feature can only be
neutral or harmful, never load-bearing. But minute_of_day never appeared
in the top-8 importances, so I am not expecting it to move the board on
its own.

The feature cache invalidates on column names, not dtypes, so this one
needs deleting by hand once the running fold job releases it.

+++ prc/features.py
@@ -331,7 +331,10 @@ def build(frame: pl.DataFrame, with_target: bool = True) -> pl.DataFrame:
     mvt = pl.col("MVT_TIME_UTC_mvt")
     frame = frame.filter(pl.col("PHASE_mvt") == "DEP").with_columns(
         mvt.dt.hour().alias("hour"),
-        (mvt.dt.hour() * 60 + mvt.dt.minute()).alias("minute_of_day"),
+        # .dt.hour() is Int8 and polars keeps that dtype through the multiply,
+        # so hour*60 overflows above 127 and wraps: 21:10 became -10, not 1270.
+        # 99.7% of rows carried a wrong value; every model so far trained on it.
+        (mvt.dt.hour().cast(pl.Int32) * 60 + mvt.dt.minute()).alias("minute_of_day"),
         mvt.dt.weekday().alias("dow"),
         mvt.dt.month().alias("month"),
         mvt.dt.ordinal_day().alias("doy"),
